New options to support using Hiera for masterless Puppet
Adding options to specify hieradata tarball.  Because using the hiera_config argument with a 'puppet apply' is only supported in Puppet > 3.1, I've also added options to specify the command to use when installing Puppet and Hiera.  If the Puppet version of your distro isn't new enough, use these options to make sure the version being installed are new enough.